pyproject.toml
[tool.poetry]
name = "mqtt2aprs"
version = "0.1.0"
description = ""
authors = ["Russ Garrett <russ@garrett.co.uk>"]

[tool.poetry.dependencies]
python = "^3.9"
aprs = "^7.0.0"
paho-mqtt = "^1.6.1"
click = "^8.1.3"
Pint = "^0.19.2"

[tool.poetry.dev-dependencies]
mypy = "^0.961"
black = {version = "^22.6.0", allow-prereleases = true}
pytest = "^7.1.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
